refactor: Modify package structure
