+++ src/main/java/racingcar/racing/Car.java
@@ -1,4 +1,4 @@
-package racingcar.car;
+package racingcar.racing;
 
 
 import racingcar.movepolicy.MoveStrategy;
