+++ src/test/java/racingcar/racing/CarTest.java
@@ -1,4 +1,4 @@
-package racingcar.car;
+package racingcar.racing;
 
 import org.junit.jupiter.api.Test;
 import racingcar.movepolicy.MoveStrategy;
